refactor(bigquery): introduce query mod
