+++ src/bigquery-derive/src/lib.rs
@@ -23,6 +23,8 @@ use syn::{Data, DeriveInput, Fields, parse_macro_input};
 /// Derives standard library [TryFrom] for converting a BigQuery `Row` into a struct.
 ///
 /// Supports renaming attributes via `#[bigquery(rename = "new_name")]`.
+///
+/// [TryFrom]: std::convert::TryFrom
 #[proc_macro_derive(FromRow, attributes(bigquery))]
 pub fn derive_from_row(input: TokenStream) -> TokenStream {
     let input = parse_macro_input!(input as DeriveInput);
@@ -61,10 +63,10 @@ pub fn derive_from_row(input: TokenStream) -> TokenStream {
     // TODO(#5592): check that the schema and this struct have same columns/attributes count.
 
     let expanded = quote! {
-        impl std::convert::TryFrom<google_cloud_bigquery::Row> for #name {
-            type Error = google_cloud_bigquery::RowError;
+        impl std::convert::TryFrom<google_cloud_bigquery::query::Row> for #name {
+            type Error = google_cloud_bigquery::error::RowError;
 
-            fn try_from(mut row: google_cloud_bigquery::Row) -> std::result::Result<Self, Self::Error> {
+            fn try_from(mut row: google_cloud_bigquery::query::Row) -> std::result::Result<Self, Self::Error> {
                 #( #value_extractions )*
 
                 std::result::Result::Ok(Self {
@@ -116,8 +118,8 @@ pub fn derive_from_sql(input: TokenStream) -> TokenStream {
         let db_column_name = get_field_name(f);
         quote! {
             let #field_name = iter.next()
-                .ok_or_else(|| google_cloud_bigquery::ConvertError::MissingField(#db_column_name.to_string()))?;
-            let #field_name = google_cloud_bigquery::FromSql::from_sql(#field_name)?;
+                .ok_or_else(|| google_cloud_bigquery::error::ConvertError::MissingField(#db_column_name.to_string()))?;
+            let #field_name = google_cloud_bigquery::query::FromSql::from_sql(#field_name)?;
         }
     });
 
@@ -126,14 +128,14 @@ pub fn derive_from_sql(input: TokenStream) -> TokenStream {
         let db_column_name = get_field_name(f);
         quote! {
             let #field_name = obj.remove(#db_column_name)
-                .ok_or_else(|| google_cloud_bigquery::ConvertError::MissingField(#db_column_name.to_string()))?;
-            let #field_name = google_cloud_bigquery::FromSql::from_sql(#field_name)?;
+                .ok_or_else(|| google_cloud_bigquery::error::ConvertError::MissingField(#db_column_name.to_string()))?;
+            let #field_name = google_cloud_bigquery::query::FromSql::from_sql(#field_name)?;
         }
     });
 
     let expanded = quote! {
-        impl google_cloud_bigquery::FromSql for #name {
-            fn from_sql(value: wkt::Value) -> std::result::Result<Self, google_cloud_bigquery::ConvertError> {
+        impl google_cloud_bigquery::query::FromSql for #name {
+            fn from_sql(value: wkt::Value) -> std::result::Result<Self, google_cloud_bigquery::error::ConvertError> {
                 match value {
                     wkt::Value::Array(arr) => {
                         let mut iter = arr.into_iter();
@@ -148,7 +150,7 @@ pub fn derive_from_sql(input: TokenStream) -> TokenStream {
                             #( #field_idents_struct_obj, )*
                         })
                     }
-                    other => std::result::Result::Err(google_cloud_bigquery::ConvertError::TypeMismatch {
+                    other => std::result::Result::Err(google_cloud_bigquery::error::ConvertError::TypeMismatch {
                         expected: "array or object",
                         got: other,
                     }),
