+++ src/bigquery/examples/src/query/params_structs.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 // [START bigquery_query_params_structs]
-use google_cloud_bigquery::FromRow;
+use google_cloud_bigquery::query::FromRow;
 use google_cloud_bigquery::client::BigQuery;
 use google_cloud_bigquery::model::{
     QueryParameter, QueryParameterStructType, QueryParameterType, QueryParameterValue,
