+++ src/bigquery/src/query/client_builder.rs
