+++ tests/bigquery/src/writes.rs
@@ -16,8 +16,8 @@ mod arrow;
 
 use crate::dataset::{cleanup_stale_datasets, create_dataset, delete_dataset, random_dataset_id};
 use anyhow::Result;
-use google_cloud_bigquery::FromRow;
 use google_cloud_bigquery::client::BigQuery;
+use google_cloud_bigquery::query::FromRow;
 use google_cloud_bigquery_v2::client::{DatasetService, TableService};
 use google_cloud_bigquery_v2::model::{Table, TableFieldSchema, TableReference, TableSchema};
 use google_cloud_bigquery_write::client::Write;
